move generated 'metrics_schema' pkg to top level

+++ scripts/build.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 set -e
-export PATH="${GOPATH}/bin:${PATH}"
-export CGO_ENABLED=0
 
 COMPONENT="$1"
 PLATFORM=$(uname | tr [:upper:] [:lower:])
@@ -9,14 +7,19 @@ GIT_REF=$(git describe --always)
 SOURCE_DIR=$(git rev-parse --show-toplevel)
 BUILD_DIR="${SOURCE_DIR}/build/${COMPONENT}"
 
-# Bring in dependencies required by schema/generator.go
-go get -u github.com/antonholmquist/jason
+if [[ $COMPONENT == "collector" ]]; then
+    export PATH="${GOPATH}/bin:${PATH}"
+    export CGO_ENABLED=0
+    rm -rf $BUILD_DIR && mkdir -p $BUILD_DIR
 
-rm -rf $BUILD_DIR && mkdir -p $BUILD_DIR
-cd "${SOURCE_DIR}/${COMPONENT}"
+    # build metrics_schema package
+    pushd "${SOURCE_DIR}/schema"
+    go run generator.go -infile metrics.avsc -outfile ./metrics_schema/schema.go
+    popd
 
-# build metrics_schema package
-go generate
-
-# build binary
-go build -a -o ${BUILD_DIR}/dcos-metrics-${COMPONENT}-${GIT_REF} *.go
+    # build binary
+    go build -a -o ${BUILD_DIR}/dcos-metrics-${COMPONENT}-${GIT_REF} *.go
+else
+    echo "Error: don't know how to build component '${COMPONENT}'!'"
+    exit 1
+fi
